requirements.txt
pyserial==3.5
Flask-SocketIO==5.3.6